feat: enhance policy repo fixture with debugger wait and execution pause logic
